requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.29
Flask-WTF==0.15.1
WTForms==2.3.3
PyMySQL==1.1.0
flask_cors==4.0.0
Flask-Admin==1.6.1
python-dotenv==1.0.1
Flask-Migrate
